Fix exception from zero thread id (#262)

+++ spark-common/src/main/java/me/lucko/spark/common/sampler/async/AsyncProfilerJob.java
@@ -249,7 +249,7 @@ public class AsyncProfilerJob {
                 duration = TimeUnit.NANOSECONDS.toMicros(sample.time - samples.get(i - 1).time);
             }
 
-            String threadName = reader.threads.get(sample.tid);
+            String threadName = reader.threads.get((long) sample.tid);
             if (threadName == null) {
                 continue;
             }
